+++ src/content-scripts/sidebar.js
@@ -336,7 +336,7 @@ async function main() {
       <div class="fixed inset-0 z-10 overflow-y-auto">
         <div class="flex items-center justify-center min-h-screen px-4 pt-4 pb-20 text-center sm:block sm:p-0">
           <div class="hidden sm:inline-block sm:h-screen sm:align-middle" aria-hidden="true"></div>
-          <div class="dark:bg-gray-900 dark:text-gray-200 dark:border-netural-400 inline-block max-h-[400px] transform overflow-hidden rounded-lg border border-gray-300 bg-white px-4 pt-5 pb-4 text-left align-bottom shadow-xl transition-all dark:bg-[#202123] sm:my-8 sm:max-h-[600px] sm:w-full sm:max-w-lg sm:p-6 sm:align-middle" role="dialog">
+          <div style="width: 80%;" class="dark:bg-gray-900 dark:text-gray-200 dark:border-netural-400 inline-block max-h-[400px] transform overflow-hidden rounded-lg border border-gray-300 bg-white px-4 pt-5 pb-4 text-left align-bottom shadow-xl transition-all dark:bg-[#202123] sm:my-8 sm:max-h-[600px] sm:w-full sm:max-w-lg sm:p-6 sm:align-middle" role="dialog">
             <div class="text-sm font-bold text-black dark:text-gray-200">${tr("name", t)}</div>
             <input style="border-color: #8e8ea0" id="prompt-name" value="${name}" class="my-2 w-full rounded-lg border border-neutral-500 px-4 py-2 text-neutral-900 shadow focus:outline-none dark:border-neutral-800 dark:border-opacity-50 dark:bg-gray-800 dark:text-neutral-100" placeholder="${tr("prompt_name", t)}" value="">
             <div class="my-2 text-sm font-bold text-black dark:text-gray-200">${tr("prompt", t)}</div>
@@ -664,8 +664,15 @@ async function main() {
             }
         }
     );
+    function removeFakeDiv(){ // convenient to place here - this prevents some themes from breaking
+        const annoyingDiv = document.querySelector(".absolute.z-10.hidden.flex-col.gap-2.right-3.top-3")
+        if (annoyingDiv) {
+            annoyingDiv.remove()
+        }
+    }
+    removeFakeDiv()
 }
-main()
+setTimeout(main, 500)
 
 function svg(name){
     switch(name){